Rename middleware files to camelCase
