+++ src/app.ts
@@ -14,8 +14,8 @@ import { registerOrdersChannel } from './bot/channels/ordersChannel';
 import { registerPaymentModerationQueue } from './bot/moderation/paymentQueue';
 import { registerVerificationModerationQueue } from './bot/moderation/verifyQueue';
 import { auth } from './bot/middlewares/auth';
-import { autoDelete } from './bot/middlewares/auto-delete';
-import { errorBoundary } from './bot/middlewares/error-boundary';
+import { autoDelete } from './bot/middlewares/autoDelete';
+import { errorBoundary } from './bot/middlewares/errorBoundary';
 import { session } from './bot/middlewares/session';
 import type { BotContext } from './bot/types';
 import { config, logger } from './config';
