+++ src/bot/middlewares/autoDelete.ts
