+++ src/bot/middlewares/errorBoundary.ts
